remove content from /etc/machine-id
In case we create more than one VM from a template
that will be added as hypervisors to same RHV engine, we will only
succeed to add one of the hypervisors becuase the other will fail due to
that both have same machine-id.

This change should remove content from /etc/machine-id and we won't bump
into this issue.